documentation-mcp/requirements.txt
mcp>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
openai>=1.3.0
numpy>=1.26.0
pyyaml>=6.0.1
lxml>=4.9.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
markdownify>=0.11.6
readability-lxml>=0.8.1
duckduckgo_search>=5.3.0
rich>=13.0.0
